% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/check_availability.R
\name{check_availability}
\alias{check_availability}
\title{Check if list items are empty}
\usage{
check_availability(list_data)
}
\arguments{
\item{list_data}{Named list of data frames}
}
\value{
Possibly empty HTML text, listing names of required data frames.
}
\description{
The function argument is a list of data frames that are required for some
subsequent analysis. If any of the data frames are empty (i.e. the data
have not been uploaded to the app), a message is returned as HTML code
listing the names of the required data frames (= names of list items).
}
\details{
In the ADAPT app, individual analyses can only be performed if the
necessary output files are uploaded. If certain files are not uploaded,
the app produces empty data frames. In the app, the function
\code{check_availability()} is used with the necessary data frames to check if they
have been uploaded and displays a message of the form "To generate this
analysis, please upload the following reports: x, y" otherwise.
}
\author{
Ian Kopacka
}
