% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_p_value.R
\name{compute_p_value}
\alias{compute_p_value}
\title{Compute p value for freedom from disease sample}
\usage{
compute_p_value(N, n, n_dis)
}
\arguments{
\item{N}{Integer; size of the population}

\item{n}{Integer; size of the sample}

\item{n_dis}{Integer; number of positives in the population}
}
\value{
Returns the probability of not finding any positives in the sample
as a numeric between 0 and 1
}
\description{
Compute the probability of drawing no positives in a sample of n items
from a Population of N containing n_dis positives.
}
\details{
The probability is computed using the hypergeometric distribution. This
function is used in \code{\link{compute_sample_size}}.
}
\seealso{
\code{\link{compute_sample_size}}
}
\author{
Ian Kopacka
}
