% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/compute_sample_size.R
\name{compute_sample_size}
\alias{compute_sample_size}
\title{Compute sample size for freedom from disease}
\usage{
compute_sample_size(N, prev, accuracy)
}
\arguments{
\item{N}{Integer containing the Size of the population}

\item{prev}{Numeric between 0 and 1; design prevalence}

\item{accuracy}{Numeric between 0 and 1; accuracy of the survey (i.e.
detection probability)}
}
\value{
Sample size (integer).
}
\description{
Compute sample size for a one stage freedom from disease survey for given
Population size, design prevalence and accuracy, assuming a perfect
diagnostic test.
}
\details{
The function finds the optimal sample size using a bisection method.
}
\author{
Ian Kopacka
}
