% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/discumulate_data.R
\name{discumulate_data}
\alias{discumulate_data}
\title{Inverse of Cumulative Sum}
\usage{
discumulate_data(value_cum)
}
\arguments{
\item{value_cum}{numeric vector; usually the result of cumulating values.}
}
\value{
A vector of the same length as \code{value_cum}
}
\description{
Computes the inverse of the \code{cumsum} function
}
\author{
Ian Kopacka
}
