% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/format_numbers_DT.R
\name{format_numbers_DT}
\alias{format_numbers_DT}
\title{Safe wrapper for DT::formatCurrency
Wrapper for \code{DT::formatCurrency} that returns NULL when the input table
is \code{NULL} (instead of throwing an error).}
\usage{
format_numbers_DT(x, ...)
}
\arguments{
\item{x}{A table object created from \code{DT::datatable()}}

\item{...}{other arguments passed to \code{DT::formatCurrency}}
}
\value{
Behaves the same output as \code{DT::formatCurrency} except when
\code{x} is \code{NULL}. Then \code{NULL} is returned and no error is thrown.
}
\description{
Safe wrapper for DT::formatCurrency
Wrapper for \code{DT::formatCurrency} that returns NULL when the input table
is \code{NULL} (instead of throwing an error).
}
\author{
Ian Kopacka
}
