% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_distribution.R
\name{plot_distribution}
\alias{plot_distribution}
\title{Plot the distribution of a variable}
\usage{
plot_distribution(x, parameter, main = "")
}
\arguments{
\item{x}{A numeric vector}

\item{parameter}{Character to use as label of the x-axis}

\item{main}{(optional) character to use as plot title}
}
\value{
No return value. Creates a plot.
}
\description{
Creates a histogram of the value along with a horizontal boxplot above it
to show the distribution of a variable.
}
\author{
Ian Kopacka
}
