% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_time_series.R
\name{plot_time_series}
\alias{plot_time_series}
\title{Plot graph of a time series with daily error margin}
\usage{
plot_time_series(x, parameter, main = "")
}
\arguments{
\item{x}{Data frame with columns \code{day} containing the counter for the
time steps (=days), \code{q025} for the 2.5 percentiles
(i.e. the lower values for the error margin), \code{median} for the median
values (i.e. the values for the line plot) and \code{q975} for the 97.5
percentiles (i.e. the upper values for the error margin).}

\item{parameter}{Character to use as label of the y-axis}

\item{main}{(optional) character to use as plot title}
}
\value{
No return value. Creates a plot.
}
\description{
Creates a line plot with a shaded polygon showing daily error margins
(uncertainty ranges)
}
\author{
Ian Kopacka
}
