% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/run_adapt.R
\name{run_adapt}
\alias{run_adapt}
\title{Run ADAPT Shiny App}
\usage{
run_adapt()
}
\value{
no return value; starts a Shiny app
}
\description{
This function runs the Shiny App "ADAPT" to analyse 'EuFMDiS' output files.
}
\details{
Upload the relevant csv output files via the "Upload files" dialog to
trigger the analysis.
}
\examples{
if (interactive()) {
    run_adapt()
}
}
\author{
Ian Kopacka
}
